chore: reorganize /src into subfolders
